Update README.md
This is code file. I have issues with my laptop. I couldn't check-in from command. So I am directly pasting code here.  This sol file takes array of strings and concat them. Solidity doesn't have inbuild support for string functions. So we have to deal like this. If you check the above code, there is no syntax error. The error would come when you try to deploy. I have found where the issue is coming. I commented the same in the code. The issue is because of infinite loops I think. Please check how to rectify it.